renamed tip to single link payments
