+++ packages/cryptoplease/lib/app/screens/authenticated/authenticated_flow_screen.dart
@@ -15,8 +15,8 @@ import '../../../di.dart';
 import '../../../features/activities/module.dart';
 import '../../../features/backup_phrase/module.dart';
 import '../../../features/favorite_tokens/module.dart';
+import '../../../features/incoming_single_key_payments/module.dart';
 import '../../../features/incoming_split_key_payments/module.dart';
-import '../../../features/incoming_tip_payments/module.dart';
 import '../../../features/intercom/module.dart';
 import '../../../features/investments/module.dart';
 import '../../../features/outgoing_direct_payments/module.dart';
