+++ packages/cryptoplease/lib/app/screens/authenticated/routes.dart
@@ -3,8 +3,8 @@ import 'package:auto_route/empty_router_widgets.dart';
 
 import '../../../features/app_lock/routes.dart';
 import '../../../features/backup_phrase/routes.dart';
+import '../../../features/incoming_single_key_payments/routes.dart';
 import '../../../features/incoming_split_key_payments/routes.dart';
-import '../../../features/incoming_tip_payments/routes.dart';
 import '../../../features/outgoing_direct_payments/routes.dart';
 import '../../../features/outgoing_split_key_payments/routes.dart';
 import '../../../features/payment_request/routes.dart';
