+++ packages/cryptoplease/lib/app/screens/authenticated/wallet_flow/wallet_flow_screen.dart
@@ -7,7 +7,7 @@ import '../../../../core/currency.dart';
 import '../../../../core/presentation/format_amount.dart';
 import '../../../../core/tokens/token_list.dart';
 import '../../../../di.dart';
-import '../../../../features/incoming_tip_payments/widgets/extensions.dart';
+import '../../../../features/incoming_single_key_payments/widgets/extensions.dart';
 import '../../../../features/outgoing_direct_payments/widgets/extensions.dart';
 import '../../../../features/outgoing_split_key_payments/widgets/extensions.dart';
 import '../../../../features/payment_request/models/payment_request.dart';
