+++ packages/cryptoplease/lib/data/db/db.dart
@@ -5,8 +5,8 @@ import '../../core/transactions/tx_sender.dart';
 import '../../features/activities/db.dart';
 import '../../features/activities/models/transaction.dart';
 import '../../features/favorite_tokens/db.dart';
+import '../../features/incoming_single_key_payments/db.dart';
 import '../../features/incoming_split_key_payments/db.dart';
-import '../../features/incoming_tip_payments/db.dart';
 import '../../features/outgoing_direct_payments/db.dart';
 import '../../features/outgoing_split_key_payments/db.dart';
 import '../../features/payment_request/db.dart';
