+++ packages/cryptoplease/lib/features/incoming_single_key_payments/src/widgets/incoming_tip_screen.dart
